chore: minimize the header files
