+++ src/scales/celsius/celsius_description.c
@@ -1,4 +1,4 @@
-#include "../utils/utils.h"
+#include "../../utils/utils.h"
 
 void celsius_description(void) {
   add_new_tab(1);
