--- src/scales/celsius_description.h
@@ -1,7 +1,0 @@
-#ifndef CELSIUS_DESCRIPTION_H
-#define CELSIUS_DESCRIPTION_H
-#include "celcius_description.c"
-
-extern void celcius_description(void);
-
-#endif
