+++ src/scales/fahrenheit/fahrenheit.h
@@ -1,0 +1,7 @@
+#ifndef FAHRENHEIT_H
+#define FAHRENHEIT_H
+#include "fahrenheit_option.c"
+
+void fahrenheit_option(void);
+
+#endif
