+++ src/scales/fahrenheit/fahrenheit_description.c
@@ -1,11 +1,11 @@
-#include "../utils/utils.h"
+#include "../../utils/utils.h"
 
-void rankine_description(void) {
+void fahrenheit_description(void) {
   add_new_tab(1);
   printf("%s--------------------------------------------------%s", BLUE, RESET);
   add_new_line(2);
   add_new_tab(1);
-  printf("    %sRankine (°R) to Other Scale%s", GREEN, RESET);
+  printf("  %sFahrenheit (°F) to Other Scale%s", GREEN, RESET);
   add_new_line(2);
   add_new_tab(1);
   printf("%s--------------------------------------------------%s", BLUE, RESET);
