--- src/scales/fahrenheit_description.h
@@ -1,7 +1,0 @@
-#ifndef FAHRENHEIT_DESCRIPTION_H
-#define FAHRENHEIT_DESCRIPTION_H
-#include "fahrenheit_description.c"
-
-extern void fahrenheit_description(void);
-
-#endif
