--- src/scales/fahrenheit_option.h
@@ -1,7 +1,0 @@
-#ifndef FAHRENHEIT_OPTION_H
-#define FAHRENHEIT_OPTION_H
-#include "fahrenheit_option.c"
-
-extern void fahrenheit_option(void);
-
-#endif
