+++ src/scales/kelvin/kelvin_description.c
@@ -1,4 +1,4 @@
-#include "../utils/utils.h"
+#include "../../utils/utils.h"
 
 void kelvin_description(void) {
   add_new_tab(1);
