--- src/scales/kelvin_description.h
@@ -1,7 +1,0 @@
-#ifndef KELVIN_DESCRIPTION_H
-#define KELVIN_DESCRIPTION_H
-#include "kelvin_description.c"
-
-extern void kelvin_description(void);
-
-#endif
