+++ src/scales/rankine/rankine.h
@@ -1,5 +1,5 @@
-#ifndef RANKINE_OPTION_H
-#define RANKINE_OPTION_H
+#ifndef RANKINE_H
+#define RANKINE_H
 #include "rankine_option.c"
 
 extern void rankine_option(void);
