--- src/scales/rankine_description.h
@@ -1,7 +1,0 @@
-#ifndef RANKINE_DESCRIPTION_H
-#define RANKINE_DESCRIPTION_H
-#include "rankine_description.c"
-
-extern void rankine_description(void);
-
-#endif
